S-01023 Data acquisition: IMU sensor: 1. added magnetometer calibration; 2. fixed a bug related to wrong raw magnetometer readings; 3. disabled logging in the device classes.

+++ src/devices/L3G.h
@@ -61,7 +61,7 @@ class L3G
 		/** Raw data holder structure. */
 		typedef struct RawData
 		{
-			float x, y, z;
+			int x, y, z;
 		} RawData;
 
 		/** Turns on and configures L3G gyroscope. */
